add fragments to drawer menu and invite to group page
