+++ android/app/src/main/java/com/example/huntertalk/InviteToGroupFragment.java
@@ -1,0 +1,21 @@
+package com.example.huntertalk;
+
+import android.os.Bundle;
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+public class InviteToGroupFragment extends Fragment {
+
+    View myView;
+
+    @Nullable
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        myView = inflater.inflate(R.layout.invite_to_group_layout, container, false);
+        return myView;
+    }
+}
