+++ android/app/src/main/res/layout/content_inside_group.xml
@@ -9,43 +9,9 @@
     tools:context=".InsideGroupActivity"
     tools:showIn="@layout/app_bar_inside_group">
 
-    <LinearLayout
-        android:layout_width="300dp"
-        android:layout_height="wrap_content"
-        android:orientation="vertical">
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="30dp"
-            android:background="@color/colorPrimary"
-            android:gravity="center"
-            android:layout_marginTop="50dp"
-            >
-            <TextView
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:text="@string/title_show_group_members"
-                android:textSize="20dp"
-                android:layout_marginLeft="5dp"/>
-
-        </LinearLayout>
-
-        <ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-            android:layout_width="match_parent"
-            android:layout_height="150dp"
-            android:fillViewport="true">
-
-            <TableLayout
-                android:id="@+id/tableGroupMembers"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:layout_marginTop="5dp"
-                android:stretchColumns="1">
-
-
-            </TableLayout>
-        </ScrollView>
-
-
-    </LinearLayout>
+    <FrameLayout
+        android:id="@+id/content_frame"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"></FrameLayout>
 
 </LinearLayout>
